Delete old HabitAdapter
